// src/dorc-web/src/apis/dorc-api/models/CloudResourceApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface CloudResourceApiModel
 */
export interface CloudResourceApiModel {
    /**
     * @type {Array<string>}
     * @memberof CloudResourceApiModel
     */
    EnvironmentNames?: Array<string> | null;
    /**
     * @type {boolean}
     * @memberof CloudResourceApiModel
     */
    UserEditable?: boolean;
    /**
     * @type {number}
     * @memberof CloudResourceApiModel
     */
    Id?: number;
    /**
     * @type {string}
     * @memberof CloudResourceApiModel
     */
    Name?: string | null;
    /**
     * @type {string}
     * @memberof CloudResourceApiModel
     */
    Provider?: string | null;
    /**
     * @type {string}
     * @memberof CloudResourceApiModel
     */
    ResourceType?: string | null;
    /**
     * @type {string}
     * @memberof CloudResourceApiModel
     */
    ResourceIdentifier?: string | null;
    /**
     * @type {string}
     * @memberof CloudResourceApiModel
     */
    Subscription?: string | null;
    /**
     * @type {string}
     * @memberof CloudResourceApiModel
     */
    Tags?: string | null;
}